inline interface
